# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A nice kanban board with a beautiful heart ✨",
        "title": "Goroutine kanban API",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/v1/boards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all boards for the current user. Results are returned in increasing creation time order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "List all boards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.boardResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new board for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Create a new board",
                "parameters": [
                    {
                        "description": "Board details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createBoardBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.boardResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "413": {
                        "description": "PAYLOAD_TOO_LARGE",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            }
        },
        "/v1/boards/{boardId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get board metadata by id for the current user (owner only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Get a board by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.boardResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "404": {
                        "description": "BOARD_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete a board and its columns and tasks for the current user (owner only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Delete a board by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "404": {
                        "description": "BOARD_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially update board metadata for the current user (owner only). Provided fields are updated; omitted or null fields are ignored.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Update a board by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Board fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateBoardBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.boardResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "404": {
                        "description": "BOARD_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "413": {
                        "description": "PAYLOAD_TOO_LARGE",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            }
        },
        "/v1/boards/{boardId}/aggregate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a board with nested columns and tasks for the current user (owner only). Columns are returned in increasing position order, and tasks inside each column are returned in increasing position order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Get a board aggregate by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Board ID",
                        "name": "boardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.aggregateBoardResponse"
                        }
                    },
                    "400": {
                        "description": "VALIDATION_ERROR",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "404": {
                        "description": "BOARD_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/httpschema.DetailedError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpschema.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.aggregateBoardResponse": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.aggregateColumnResponse"
                    }
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                },
                "description": {
                    "type": "string",
                    "example": "My Todo Description"
                },
                "id": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a2"
                },
                "name": {
                    "type": "string",
                    "example": "My Todo Name"
                },
                "ownerId": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a2"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                }
            }
        },
        "handler.aggregateColumnResponse": {
            "type": "object",
            "properties": {
                "boardId": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a1"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                },
                "description": {
                    "type": "string",
                    "example": "My Column Description"
                },
                "id": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a2"
                },
                "name": {
                    "type": "string",
                    "example": "In Progress"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.taskResponse"
                    }
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                }
            }
        },
        "handler.boardResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                },
                "description": {
                    "type": "string",
                    "example": "My Todo Description"
                },
                "id": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a2"
                },
                "name": {
                    "type": "string",
                    "example": "My Todo Name"
                },
                "ownerId": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a2"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                }
            }
        },
...
        "handler.createBoardBody": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "My Board Description"
                },
                "name": {
                    "type": "string",
                    "example": "My Board Name"
                }
            }
        },
...
        "handler.taskResponse": {
            "type": "object",
            "properties": {
                "columnId": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a2"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                },
                "description": {
                    "type": "string",
                    "example": "Cover the new endpoint with tests"
                },
                "id": {
                    "type": "string",
                    "example": "019cc971-e5be-7df9-ae8a-c6e3f29c86a3"
                },
                "name": {
                    "type": "string",
                    "example": "Write tests"
                },
                "position": {
                    "type": "integer",
                    "example": 1
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-03-07T20:56:50.000+03:00"
                }
            }
        },
...
        "handler.updateBoardBody": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "My Board Description"
                },
                "name": {
                    "type": "string",
                    "example": "My Board Name"
                }
            }
        },
...
        "httpschema.Detail": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "email"
                },
                "issues": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "must be a valid email",
                        "too short"
                    ]
                }
            }
        },
        "httpschema.DetailedError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_CREDENTIALS"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpschema.Detail"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Invalid login or password"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-03-02T15:04:05.123Z"
                }
            }
        },
        "httpschema.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_CREDENTIALS"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid login or password"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-03-02T15:04:05.123Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token with `Bearer ` prefix, e.g. `Bearer eyJhbGciOi...`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}