# internal/http/handler/boards.go
package handler

import (
	"context"
	"errors"
	"log/slog"
	"net/http"

	"goroutine/internal/domain"
	"goroutine/internal/http/httpschema"
	"goroutine/internal/logging"
	"goroutine/internal/service"
)

type BoardsService interface {
	Create(ctx context.Context, ownerID domain.UserID, name domain.BoardName, description domain.BoardDescription) (domain.Board, error)
	Get(ctx context.Context, ownerID domain.UserID, boardID domain.BoardID) (domain.Board, error)
	GetAggregate(ctx context.Context, ownerID domain.UserID, boardID domain.BoardID) (service.AggregateBoard, error)
	ListByOwnerID(ctx context.Context, ownerID domain.UserID) ([]domain.Board, error)
	Update(ctx context.Context, ownerID domain.UserID, boardID domain.BoardID, name *domain.BoardName, description *domain.BoardDescription) (domain.Board, error)
	Delete(ctx context.Context, ownerID domain.UserID, boardID domain.BoardID) error
}

type Boards struct {
	logger        *slog.Logger
	boardsService BoardsService
	responder     *httpschema.ErrorResponder
}

func NewBoards(logger *slog.Logger, boardsService BoardsService, responder *httpschema.ErrorResponder) *Boards {
	moduleLogger := logging.WithModule(logger, "handler.boards")

	return &Boards{logger: moduleLogger, boardsService: boardsService, responder: responder}
}

type createBoardBody struct {
	Name        string `json:"name" example:"My Board Name"`
	Description string `json:"description" example:"My Board Description"`
}

type updateBoardBody struct {
	Name        *string `json:"name" example:"My Board Name"`
	Description *string `json:"description" example:"My Board Description"`
}

type boardResponse struct {
	ID          string `json:"id" example:"019cc971-e5be-7df9-ae8a-c6e3f29c86a2"`
	OwnerID     string `json:"ownerId" example:"019cc971-e5be-7df9-ae8a-c6e3f29c86a2"`
	Name        string `json:"name" example:"My Todo Name"`
	Description string `json:"description" example:"My Todo Description"`
	CreatedAt   string `json:"createdAt" example:"2026-03-07T20:56:50.000+03:00"`
	UpdatedAt   string `json:"updatedAt" example:"2026-03-07T20:56:50.000+03:00"`
}

func newBoardResponse(board *domain.Board) boardResponse {
	return boardResponse{
		ID:          board.ID.String(),
		OwnerID:     board.OwnerID.String(),
		Name:        board.Name.String(),
		Description: board.Description.String(),

		CreatedAt: service.FormatRFC3339Millis(board.CreatedAt),
		UpdatedAt: service.FormatRFC3339Millis(board.UpdatedAt),
	}
}

type aggregateBoardResponse struct {
	boardResponse
	Columns []aggregateColumnResponse `json:"columns"`
}

type aggregateColumnResponse struct {
	columnResponse
	Tasks []taskResponse `json:"tasks"`
}

func newBoardAggregateResponse(aggregateBoard *service.AggregateBoard) aggregateBoardResponse {
	columnResps := make([]aggregateColumnResponse, len(aggregateBoard.Columns))

	for i := range aggregateBoard.Columns {
		column := &aggregateBoard.Columns[i]

		taskResps := make([]taskResponse, len(column.Tasks))
		for j := range column.Tasks {
			taskResps[j] = newTaskResponse(&column.Tasks[j])
		}

		columnResps[i] = aggregateColumnResponse{
			columnResponse: newColumnResponse(&column.Column),
			Tasks:          taskResps,
		}
	}

	return aggregateBoardResponse{
		boardResponse: newBoardResponse(&aggregateBoard.Board),
		Columns:       columnResps,
	}
}

type listBoardsResponse = []boardResponse

// Create godoc
// @Summary Create a new board
// @Description Create a new board for the current user
// @Tags boards
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param body body createBoardBody true "Board details"
// @Success 201 {object} boardResponse
// @Failure 400 {object} httpschema.DetailedError "VALIDATION_ERROR"
// @Failure 413 {object} httpschema.DetailedError "PAYLOAD_TOO_LARGE"
// @Failure 401 {object} httpschema.DetailedError "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER"
// @Failure 500 {object} httpschema.Error "Internal server error"
// @Router /v1/boards [post]
func (h *Boards) Create(w http.ResponseWriter, r *http.Request) {
	var body createBoardBody

	err := decodeJSONLimited(r, &body)
	if err != nil {
		if errors.Is(err, errBodyTooLarge) {
			h.responder.PayloadTooLarge(w)
		} else {
			h.responder.ValidationError(w, []httpschema.Detail{{Field: "body", Issues: []string{"Invalid JSON body"}}})
		}
		return
	}

	details := []httpschema.Detail{}
	name := httpschema.ValidateField("name", body.Name, domain.NewBoardName, &details)
	description := httpschema.ValidateField("description", body.Description, domain.NewBoardDescription, &details)
	if len(details) > 0 {
		h.responder.ValidationError(w, details)
		return
	}

	userID, ok := extractUserIDOrHandleMissing(w, r, h.logger, h.responder)
	if !ok {
		return
	}

	board, err := h.boardsService.Create(r.Context(), userID, name, description)
	if err != nil {
		h.responder.InternalError(w, r, err)
		return
	}

	httpschema.RespondJSON(w, h.logger, http.StatusCreated, newBoardResponse(&board))
}

// Get godoc
// @Summary Get a board by id
// @Description Get board metadata by id for the current user (owner only)
// @Tags boards
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param boardId path string true "Board ID"
// @Success 200 {object} boardResponse
// @Failure 400 {object} httpschema.DetailedError "VALIDATION_ERROR"
// @Failure 401 {object} httpschema.DetailedError "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER"
// @Failure 404 {object} httpschema.DetailedError "BOARD_NOT_FOUND"
// @Failure 500 {object} httpschema.Error "Internal server error"
// @Router /v1/boards/{boardId} [get]
func (h *Boards) Get(w http.ResponseWriter, r *http.Request) {
	rawID := r.PathValue("boardId")
	boardID, err := domain.ParseBoardID(rawID)
	if err != nil {
		h.responder.ValidationError(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Invalid board id"}}})
		return
	}

	userID, ok := extractUserIDOrHandleMissing(w, r, h.logger, h.responder)
	if !ok {
		return
	}

	board, err := h.boardsService.Get(r.Context(), userID, boardID)
	if err != nil {
		if errors.Is(err, service.ErrBoardNotFound) {
			h.responder.BoardNotFound(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Board not found"}}})
			return
		}
		h.responder.InternalError(w, r, err)
		return
	}

	httpschema.RespondJSON(w, h.logger, http.StatusOK, newBoardResponse(&board))
}

// GetAggregate godoc
// @Summary Get a board aggregate by id
// @Description Get a board with nested columns and tasks for the current user (owner only). Columns are returned in increasing position order, and tasks inside each column are returned in increasing position order.
// @Tags boards
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param boardId path string true "Board ID"
// @Success 200 {object} aggregateBoardResponse
// @Failure 400 {object} httpschema.DetailedError "VALIDATION_ERROR"
// @Failure 401 {object} httpschema.DetailedError "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER"
// @Failure 404 {object} httpschema.DetailedError "BOARD_NOT_FOUND"
// @Failure 500 {object} httpschema.Error "Internal server error"
// @Router /v1/boards/{boardId}/aggregate [get]
func (h *Boards) GetAggregate(w http.ResponseWriter, r *http.Request) {
	rawID := r.PathValue("boardId")
	boardID, err := domain.ParseBoardID(rawID)
	if err != nil {
		h.responder.ValidationError(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Invalid board id"}}})
		return
	}

	userID, ok := extractUserIDOrHandleMissing(w, r, h.logger, h.responder)
	if !ok {
		return
	}

	aggregate, err := h.boardsService.GetAggregate(r.Context(), userID, boardID)
	if err != nil {
		if errors.Is(err, service.ErrBoardNotFound) {
			h.responder.BoardNotFound(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Board not found"}}})
			return
		}
		h.responder.InternalError(w, r, err)
		return
	}

	httpschema.RespondJSON(w, h.logger, http.StatusOK, newBoardAggregateResponse(&aggregate))
}

// List godoc
// @Summary List all boards
// @Description List all boards for the current user. Results are returned in increasing creation time order.
// @Tags boards
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} listBoardsResponse
// @Failure 400 {object} httpschema.DetailedError "VALIDATION_ERROR"
// @Failure 401 {object} httpschema.DetailedError "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER"
// @Failure 500 {object} httpschema.Error "Internal server error"
// @Router /v1/boards [get]
func (h *Boards) ListByOwnerID(w http.ResponseWriter, r *http.Request) {
	userID, ok := extractUserIDOrHandleMissing(w, r, h.logger, h.responder)
	if !ok {
		return
	}

	boards, err := h.boardsService.ListByOwnerID(r.Context(), userID)
	if err != nil {
		h.responder.InternalError(w, r, err)
		return
	}

	response := make(listBoardsResponse, len(boards))
	for i := range boards {
		response[i] = newBoardResponse(&boards[i])
	}

	httpschema.RespondJSON(w, h.logger, http.StatusOK, response)
}

// Update godoc
// @Summary Update a board by id
// @Description Partially update board metadata for the current user (owner only). Provided fields are updated; omitted or null fields are ignored.
// @Tags boards
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param boardId path string true "Board ID"
// @Param body body updateBoardBody true "Board fields to update"
// @Success 200 {object} boardResponse
// @Failure 400 {object} httpschema.DetailedError "VALIDATION_ERROR"
// @Failure 413 {object} httpschema.DetailedError "PAYLOAD_TOO_LARGE"
// @Failure 401 {object} httpschema.DetailedError "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER"
// @Failure 404 {object} httpschema.DetailedError "BOARD_NOT_FOUND"
// @Failure 500 {object} httpschema.Error "Internal server error"
// @Router /v1/boards/{boardId} [patch]
func (h *Boards) Update(w http.ResponseWriter, r *http.Request) {
	rawID := r.PathValue("boardId")
	boardID, err := domain.ParseBoardID(rawID)
	if err != nil {
		h.responder.ValidationError(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Invalid board id"}}})
		return
	}

	var body updateBoardBody
	err = decodeJSONLimited(r, &body)
	if err != nil {
		if errors.Is(err, errBodyTooLarge) {
			h.responder.PayloadTooLarge(w)
		} else {
			h.responder.ValidationError(w, []httpschema.Detail{{Field: "body", Issues: []string{"Invalid JSON body"}}})
		}
		return
	}

	details := []httpschema.Detail{}
	var name *domain.BoardName
	if body.Name != nil {
		value := httpschema.ValidateField("name", *body.Name, domain.NewBoardName, &details)
		name = &value
	}

	var description *domain.BoardDescription
	if body.Description != nil {
		value := httpschema.ValidateField("description", *body.Description, domain.NewBoardDescription, &details)
		description = &value
	}

	if len(details) > 0 {
		h.responder.ValidationError(w, details)
		return
	}

	userID, ok := extractUserIDOrHandleMissing(w, r, h.logger, h.responder)
	if !ok {
		return
	}

	board, err := h.boardsService.Update(r.Context(), userID, boardID, name, description)
	if err != nil {
		if errors.Is(err, service.ErrBoardNotFound) {
			h.responder.BoardNotFound(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Board not found"}}})
			return
		}
		h.responder.InternalError(w, r, err)
		return
	}

	httpschema.RespondJSON(w, h.logger, http.StatusOK, newBoardResponse(&board))
}

// Delete godoc
// @Summary Delete a board by id
// @Description Permanently delete a board and its columns and tasks for the current user (owner only)
// @Tags boards
// @Accept json
// @Produce json
// @Security BearerAuth
// @Param boardId path string true "Board ID"
// @Success 204 "No Content"
// @Failure 400 {object} httpschema.DetailedError "VALIDATION_ERROR"
// @Failure 401 {object} httpschema.DetailedError "Unauthorized: INVALID_TOKEN or INVALID_AUTH_HEADER"
// @Failure 404 {object} httpschema.DetailedError "BOARD_NOT_FOUND"
// @Failure 500 {object} httpschema.Error "Internal server error"
// @Router /v1/boards/{boardId} [delete]
func (h *Boards) Delete(w http.ResponseWriter, r *http.Request) {
	rawID := r.PathValue("boardId")
	boardID, err := domain.ParseBoardID(rawID)
	if err != nil {
		h.responder.ValidationError(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Invalid board id"}}})
		return
	}

	userID, ok := extractUserIDOrHandleMissing(w, r, h.logger, h.responder)
	if !ok {
		return
	}

	err = h.boardsService.Delete(r.Context(), userID, boardID)
	if err != nil {
		if errors.Is(err, service.ErrBoardNotFound) {
			h.responder.BoardNotFound(w, []httpschema.Detail{{Field: "boardId", Issues: []string{"Board not found"}}})
			return
		}
		h.responder.InternalError(w, r, err)

		return
	}

	w.WriteHeader(http.StatusNoContent)
}